--- src/app/store/actions/complaint.actions.ts
@@ -1,11 +1,0 @@
-import { createAction, props } from "@ngrx/store";
-import { Complaint } from '../models/complaint.model';
-
-export enum ComplaintActionType {
-  ADD_ITEM = '[COMPLAINT] Add COMPLAINT',
-}
-
-export const AddComplaintAction = createAction(
-  ComplaintActionType.ADD_ITEM,
-  props<{payload: Complaint}>()
-);
