--- src/app/store/models/complaint.model.ts
@@ -1,8 +1,0 @@
-export interface Complaint {
-    id: string;
-    title: string,
-    userId: string,
-    description: string,
-    status: string,
-    Priority: string
-  }
