--- src/app/store/models/state.model.ts
@@ -1,5 +1,0 @@
-import { Complaint } from './complaint.model';
-
-export interface State {
-  readonly complaint: Array<Complaint>;
-}
