+++ src/store/models/index.ts
