+++ src/store/models/state.model.ts
@@ -1,0 +1,5 @@
+import { Reclamation } from './reclamation';
+
+export interface State {
+  readonly complaint: Array<Reclamation>;
+}
